fix(core): fix minGiftValue only receives integer

+++ apps/core/src/app/config/merge-query.ts
@@ -43,7 +43,7 @@ export function mergeQueryParameters(query:any,config:ComenConfiguration){
             // No other effect
         }
         if(query.minGiftValue!=undefined){
-            config.minGiftValue = parseInt(query.minGiftValue);
+            config.minGiftValue = parseFloat(query.minGiftValue);
         }
         if(query.silverGoldRatio!=undefined){
             config.silverGoldRatio = parseInt(query.silverGoldRatio);
